refactored to move all exception constants to the root package 'exception'
