+++ wls-eai-service/src/main/java/de/muenchen/oss/wahllokalsystem/eaiservice/service/IDConverter.java
@@ -1,6 +1,6 @@
 package de.muenchen.oss.wahllokalsystem.eaiservice.service;
 
-import de.muenchen.oss.wahllokalsystem.eaiservice.rest.common.exception.ExceptionConstants;
+import de.muenchen.oss.wahllokalsystem.eaiservice.exception.ExceptionConstants;
 import de.muenchen.oss.wahllokalsystem.wls.common.exception.util.ExceptionFactory;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
