+++ wls-eai-service/src/test/java/de/muenchen/oss/wahllokalsystem/eaiservice/service/IDConverterTest.java
@@ -1,6 +1,6 @@
 package de.muenchen.oss.wahllokalsystem.eaiservice.service;
 
-import de.muenchen.oss.wahllokalsystem.eaiservice.rest.common.exception.ExceptionConstants;
+import de.muenchen.oss.wahllokalsystem.eaiservice.exception.ExceptionConstants;
 import de.muenchen.oss.wahllokalsystem.wls.common.exception.FachlicheWlsException;
 import de.muenchen.oss.wahllokalsystem.wls.common.exception.util.ExceptionFactory;
 import lombok.val;
